Structure has been changed to MVC(Model, View, Controller) Pattern
