+++ Twitter/View/Cells.swift
@@ -1,0 +1,56 @@
+//
+//  Cells.swift
+//  Twitter
+//
+//  Created by Jiwoo Ban on 2020/04/25.
+//  Copyright © 2020 Jiwoo Ban. All rights reserved.
+//
+
+import LBTAComponents
+
+
+//헤더추가하기
+class UserHeader: DatasourceCell {
+    override func setupViews() {
+        super.setupViews()
+        backgroundColor = .black
+    }
+}
+
+//푸터추가하기
+class UserFooter: DatasourceCell {
+    override func setupViews() {
+        super.setupViews()
+        backgroundColor = .lightGray
+    }
+}
+
+
+//cell에 색깔을 넣어봅시다.
+class UserCell: DatasourceCell {
+    
+    
+    
+    //정의한 word들을 레이블에 넣기
+    override var datasourceItem: Any? {
+        didSet {
+            nameLabel.text = datasourceItem as? String
+        }
+    }
+    
+    let nameLabel: UILabel = {
+        let label = UILabel()
+        label.text = "T T T"
+        return label
+    }()
+    
+    override func setupViews() {
+        super.setupViews()
+        backgroundColor = .yellow
+        
+        addSubview(nameLabel)
+        nameLabel.anchor(topAnchor, left: leftAnchor, bottom: bottomAnchor, right: rightAnchor, topConstant: 10, leftConstant: 50, bottomConstant: 0, rightConstant: 0, widthConstant: 0, heightConstant: 0)
+        
+    }
+}
+
